requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
blis==0.7.9
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
dnspython==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-language==2.6.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gspread==5.7.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1